+++ terrareg/server/__init__.py
@@ -2,7 +2,7 @@
 import os
 from functools import wraps
 
-from flask import Flask, session, redirect
+from flask import Flask, session, redirect, request
 from flask_restful import Api
 import sentry_sdk
 from sentry_sdk.integrations.flask import FlaskIntegration
@@ -183,6 +183,7 @@ class Server(BaseHandler):
         )
         self._api.add_resource(
             ApiModuleVersionDownload,
+            '/v1/modules/<string:namespace>/<string:name>/<string:provider>/download',
             '/v1/modules/<string:namespace>/<string:name>/<string:provider>/<string:version>/download'
         )
         self._api.add_resource(
@@ -666,6 +667,14 @@ class Server(BaseHandler):
     @catch_name_exceptions
     def _view_serve_module_provider(self, namespace, name, provider, version=None):
         """Render view for displaying module provider information"""
+
+        # Check if used by terraform and return module download Api
+        if request.args.get("terraform-get") == "1":
+            # Convert name 'latest' for version to None
+            if version == 'latest':
+                version = None
+            return ApiModuleVersionDownload().get(namespace=namespace, name=name, provider=provider, version=version)
+
         namespace_obj = terrareg.models.Namespace.get(namespace)
         if namespace_obj is None:
             return self._namespace_404(