+++ terrareg/server/api/terrareg_example_details.py
@@ -1,4 +1,8 @@
 
+import urllib.parse
+
+from flask import request
+
 from terrareg.server.error_catching_resource import ErrorCatchingResource
 import terrareg.models
 
@@ -14,4 +18,5 @@ class ApiTerraregExampleDetails(ErrorCatchingResource):
 
         example_obj = terrareg.models.Example.get(module_version=module_version, module_path=example)
 
-        return example_obj.get_terrareg_api_details()
+        return example_obj.get_terrareg_api_details(
+            request_domain=urllib.parse.urlparse(request.base_url).hostname)