+++ terrareg/server/api/terrareg_module_version_details.py
@@ -1,4 +1,8 @@
 
+import urllib.parse
+
+from flask import request
+
 from terrareg.server.error_catching_resource import ErrorCatchingResource
 import terrareg.models
 
@@ -20,4 +24,5 @@ class ApiTerraregModuleVersionDetails(ErrorCatchingResource):
         if module_version is None:
             return self._get_404_response()
 
-        return module_version.get_terrareg_api_details()
+        return module_version.get_terrareg_api_details(
+            request_domain=urllib.parse.urlparse(request.base_url).hostname)